SEARCHBOT/requirements.txt
python-telegram-bot>=20.7
httpx[http2]>=0.27
beautifulsoup4>=4.12
ddgs>=8.0.0
playwright>=1.44
python-dotenv>=1.0
lxml>=5.0
apscheduler>=3.10
